requirements.txt
piper-tts
pynput
PySide6
PyYAML
desktop